# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Comment:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorID:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isReply:
        type: boolean
      postID:
        type: string
      replyToID:
        type: string
      reputation:
        type: integer
    type: object
  models.Group:
    properties:
      bannerURL:
        type: string
      description:
        type: string
      groupName:
        type: string
      id:
        type: string
      moderators:
        items:
          $ref: '#/definitions/models.User'
        type: array
      registeredAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      bannerURL:
        type: string
      email:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: string
      nickname:
        type: string
      passwordHash:
        type: string
      registeredAt:
        type: string
      reputationComments:
        type: integer
      reputationPosts:
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  routes.CommentDTO:
    properties:
      authorId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isReply:
        type: boolean
      postId:
        type: string
      replyToId:
        type: string
      reputation:
        type: integer
    type: object
  routes.CreateCommentRequest:
    properties:
      content:
        type: string
      postId:
        type: string
      replyToId:
        type: string
    required:
    - content
    - postId
    type: object
...
  routes.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  routes.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  routes.RefreshResponse:
    properties:
      token:
        type: string
    type: object
  routes.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - nickname
    - password
    type: object
  routes.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      registeredAt:
        type: string
    type: object
  routes.UpdateCommentDTO:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  routes.VoteDTO:
    properties:
      value:
        enum:
        - -1
        - 1
        type: integer
    required:
    - value
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Reddit-like REST API.
  title: Chirp API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя и выдача JWT
      parameters:
      - description: Данные для входа
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход пользователя
      tags:
      - auth
  /auth/refresh:
    post:
      description: Обновляет JWT токен пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить JWT токен
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
  /comments:
    post:
      consumes:
      - application/json
      description: Создаёт новый комментарий к посту
      parameters:
      - description: Данные для комментария
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.CommentDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать комментарий
      tags:
      - comments
  /comments/{id}:
    delete:
      description: Удаляет комментарий пользователя
      parameters:
      - description: ID комментария
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить комментарий
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Обновляет комментарий пользователя
      parameters:
      - description: ID комментария
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateCommentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить комментарий
      tags:
      - comments
  /comments/{id}/vote:
    post:
      consumes:
      - application/json
      description: Голосует за комментарий (лайк/дизлайк)
      parameters:
      - description: ID комментария
        in: path
        name: id
        required: true
        type: string
      - description: Голос
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.VoteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Голосовать за комментарий
      tags:
      - comments
  /comments/posts/{id}/comments:
    get:
      description: Получает все комментарии к посту
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить комментарии к посту
      tags:
      - comments
...
swagger: "2.0"